\name{bind.relative}
\alias{bind.relative}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Bind tips to their designated relatives
}
\description{
This function binds a list of species to their designated species- and genus-level relatives in the phylogeny.
}
\usage{
bind.relative(sp.list, tree, gen.list, nodes.type = 1, output.sp.list = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sp.list}{
The user-provided species list.
}
  \item{tree}{
A phylo object of the megaphylogeny, which is the backbone of making the phylogenetic hypotheses of sp.list.
}
  \item{gen.list}{
A data frame with two columns, column one includes the names of all extant genera of the species group on the phylogeny, column two includes the names of the families of the genera in column one. The file containing the information can be downloaded online.
}
  \item{nodes.type}{
The type of nodes information of the backbone phylogeny selected by the user. The default is 1, which is generated by build.nodes.1 function, the other option is 2, generated by build.nodes.2.
}
  \item{output.sp.list}{
Whether or not to output the species list, with each species' status of binded to its designated relative or not. The default is TRUE.
}
}
\details{
}
\value{

  \item{phylo }{The updated phylogeny, with species binded to the designated relative}
  \item{species.list }{The species list with the information of each species status of binded to its designated relative or not.}
  \item{nodes.info }{The nodes information of the updated phylogeny.}

}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Yi Jin
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
